+++ lib/cgo/tests/check_params.distribution.c
@@ -6,7 +6,7 @@
 #include "skyerrors.h"
 #include "skystring.h"
 #include "skytest.h"
-#include "skycriterion.h"
+#include "skyassert.h"
 
 // TestSuite(params_distribution, .init = setup, .fini = teardown);