+++ lib/cgo/tests/check_util.droplet.c
@@ -3,7 +3,7 @@
 #include <string.h>
 
 #include "libskycoin.h"
-#include "skycriterion.h"
+#include "skyassert.h"
 #include "skyerrors.h"
 #include "skystring.h"
 #include "skytest.h"