+++ lib/cgo/tests/check_util.fee.c
@@ -4,7 +4,7 @@
 
 #include "libskycoin.h"
 #include "math.h"
-#include "skycriterion.h"
+#include "skyassert.h"
 #include "skyerrors.h"
 #include "skystring.h"
 #include "skytest.h"