+++ .travis/install_lib_curl.sh
@@ -1,0 +1,10 @@
+cd lib/curl
+
+mkdir -p build
+cd build
+# for normal install use following command
+cmake ..
+make
+sudo make install
+
+cd ../..