+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+
+### Added
 - Add function `SKY_base58_Hex2Base58`
 - Add function `SKY_base58_Encode`
 - Add function `SKY_base58_Decode`
@@ -21,7 +24,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add function `SKY_httphelper_SHA256_UnmarshalJSON`
 - Add function `SKY_httphelper_SHA256_MarshalJSON`
 - Add data type `httphelper__SHA256`
-- Delete function `SKY_base58_String2Hex`
 - Add function `SKY_wallet_Entry_SkycoinAddress`
 - Add function `SKY_wallet_Entry_BitcoinAddress`
 - Add function `SKY_wallet_ResolveCoinType`
@@ -32,8 +34,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add function `SKY_wallet_Wallet_CreateAndSignTransaction`
 - Add function `CreateTransactionParams_Handle`
 - Add function `SKY_wallet_Wallet_CreateAndSignTransactionAdvanced`
-
-### Added
 - Add CLI `addressTransactions` command
 - Add `/api/v2/wallet/seed/verify` to verify if seed is a valid bip39 mnemonic seed
 
@@ -48,6 +48,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ### Removed
+- Function `SKY_base58_String2Hex`
 
 ## [0.25.0] - 2018-12-19